Remove duplication in the  localization tests
